refactor(auth): extract signup service out of the auth controller
Move the account-signup business logic (capacity gate, invite eligibility,
mass-assignment scrubbing, user creation, email verification, organization
provisioning, analytics, invite finalize/release) out of
auth.controller.signup into a new auth.signup.service.js. The controller
keeps only the response block (token/cookie/JSON) and error mapping.

isMailerConfigured/sendVerificationEmail move with it (both had callers
outside signup - getConfig and resendVerification - which now re-import
them from the service). The capacity/eligibility gate rejection is
signaled via an AppError with code SIGNUP_DISABLED so the controller
reconstructs the exact original 404 response instead of the generic 422
fallback.

Pure refactor - behaviour unchanged. Verified against a baseline run of
the unmodified auth suites (14/82 unit, 5/122 integration, 1/4 e2e); the
refactored branch reproduces the same counts.

Refs #3995
